Downloads/homeworks5-master/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-dotenv==1.4.2
django-querycount==0.8.3
djangorestframework==3.15.2
drf-yasg==1.21.9
inflection==0.5.1
packaging==24.2
psycopg==3.2.5
psycopg2==2.9.10
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
